+++ servicetalk-http-api/src/main/java/io/servicetalk/http/api/PayloadTooLargeException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.servicetalk.http.utils;
+package io.servicetalk.http.api;
 
 /**
  * Indication that the payload was too large and failed to process.
